J2534_WIN: STYLE ONLY. Split files for my sanity.
