+++ drivers/windows/.gitignore
@@ -301,3 +301,6 @@ __pycache__/
 *.btm.cs
 *.odx.cs
 *.xsd.cs
+
+# installer
+*.exe
