+++ drivers/windows/pandaJ2534DLL/stdafx.h
@@ -1,8 +1,3 @@
-// stdafx.h : include file for standard system include files,
-// or project specific include files that are used frequently, but
-// are changed infrequently
-//
-
 #pragma once
 
 #include "targetver.h"
@@ -11,5 +6,8 @@
 // Windows Header Files:
 #include <windows.h>
 
-
-// TODO: reference additional headers your program requires here
+#include <string>
+#include <array>
+#include <queue>
+#include <ctime>
+#include <chrono>
